+++ maphero/android/src/main/java/com/maphero/app/GlobalMethodHandler.java
@@ -4,12 +4,21 @@ import android.content.Context;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import org.maplibre.android.net.ConnectivityReceiver;
+
+import org.maphero.android.net.ConnectivityReceiver;
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.Map;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.PluginRegistry;
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.File;
@@ -26,15 +35,9 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
   private static final int BUFFER_SIZE = 1024 * 2;
   @NonNull private final Context context;
   @NonNull private final BinaryMessenger messenger;
-  @Nullable private PluginRegistry.Registrar registrar;
   @Nullable private FlutterPlugin.FlutterAssets flutterAssets;
   @Nullable private OfflineChannelHandlerImpl downloadOfflineRegionChannelHandler;
 
-  GlobalMethodHandler(@NonNull PluginRegistry.Registrar registrar) {
-    this.registrar = registrar;
-    this.context = registrar.activeContext();
-    this.messenger = registrar.messenger();
-  }
 
   GlobalMethodHandler(@NonNull FlutterPlugin.FlutterPluginBinding binding) {
     this.context = binding.getApplicationContext();
@@ -70,7 +73,7 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
 
   @Override
   public void onMethodCall(MethodCall methodCall, MethodChannel.Result result) {
-    MapHeroUtils.getMapLibre(context);
+    MapHeroUtils.getMapHero(context);
 
     switch (methodCall.method) {
       case "installOfflineMapTiles":
@@ -88,7 +91,7 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
         break;
       case "setOfflineTileCountLimit":
         OfflineManagerUtils.setOfflineTileCountLimit(
-            result, context, methodCall.<Number>argument("limit").longValue());
+                result, context, methodCall.<Number>argument("limit").longValue());
         break;
       case "setHttpHeaders":
         Map<String, String> headers = (Map<String, String>) methodCall.argument("headers");
@@ -116,7 +119,7 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
 
         // Start downloading
         OfflineManagerUtils.downloadRegion(
-            result, context, definitionMap, metadataMap, downloadOfflineRegionChannelHandler);
+                result, context, definitionMap, metadataMap, downloadOfflineRegionChannelHandler);
         downloadOfflineRegionChannelHandler = null;
         break;
       case "getListOfRegions":
@@ -126,11 +129,11 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
         // Get download region arguments from caller
         Map<String, Object> metadata = (Map<String, Object>) methodCall.argument("metadata");
         OfflineManagerUtils.updateRegionMetadata(
-            result, context, methodCall.<Number>argument("id").longValue(), metadata);
+                result, context, methodCall.<Number>argument("id").longValue(), metadata);
         break;
       case "deleteOfflineRegion":
         OfflineManagerUtils.deleteRegion(
-            result, context, methodCall.<Number>argument("id").longValue());
+                result, context, methodCall.<Number>argument("id").longValue());
         break;
       default:
         result.notImplemented();
@@ -141,7 +144,7 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
   private void installOfflineMapTiles(String tilesDb) {
     final File dest = new File(context.getFilesDir(), DATABASE_NAME);
     try (InputStream input = openTilesDbFile(tilesDb);
-        OutputStream output = new FileOutputStream(dest)) {
+         OutputStream output = new FileOutputStream(dest)) {
       copy(input, output);
     } catch (IOException e) {
       e.printStackTrace();
@@ -153,9 +156,7 @@ class GlobalMethodHandler implements MethodChannel.MethodCallHandler {
       return new FileInputStream(new File(tilesDb));
     } else {
       String assetKey;
-      if (registrar != null) {
-        assetKey = registrar.lookupKeyForAsset(tilesDb);
-      } else if (flutterAssets != null) {
+      if (flutterAssets != null) {
         assetKey = flutterAssets.getAssetFilePathByName(tilesDb);
       } else {
         throw new IllegalStateException();