+++ maphero/android/src/main/java/com/maphero/app/setMapLanguage.kt
@@ -2,12 +2,13 @@
 
 package com.maphero.app
 
-import org.maplibre.android.maps.MapLibreMap
-import org.maplibre.android.style.expressions.Expression
-import org.maplibre.android.style.layers.PropertyFactory
-import org.maplibre.android.style.layers.SymbolLayer
+import org.maphero.android.maps.MapHeroMap
+import org.maphero.android.style.expressions.Expression
+import org.maphero.android.style.layers.PropertyFactory
+import org.maphero.android.style.layers.SymbolLayer
 
-fun MapLibreMap.setMapLanguage(language: String) {
+
+fun MapHeroMap.setMapLanguage(language: String) {
     val layers = this.style?.layers ?: emptyList()
 
     val languageRegex = Regex("(name:[a-z]+)")