+++ maphero/CHANGELOG.md
@@ -1,3 +1,45 @@
+## 0.0.20
+  Migrate to MapHero
+
+## 0.0.19
+  Migrate to MapHero
+
+## 0.0.18
+  Migrate to MapHero
+
+## 0.0.17
+  Migrate to MapHero
+
+## 0.0.16
+  Migrate to MapHero
+
+## 0.0.15
+  Migrate to MapHero
+
+## 0.0.14
+  Migrate to MapHero
+
+## 0.0.13
+  Migrate to MapHero
+
+## 0.0.12
+  Migrate to MapHero
+
+## 0.0.11
+  Migrate to MapHero
+
+## 0.0.10
+  Migrate to MapHero
+
+## 0.0.9
+  Migrate to MapHero
+
+## 0.0.8
+  Migrate to MapHero
+
+## 0.0.7
+  Migrate to MapHero
+
 ## 0.0.5
   Migrate to MapHero